refactor StateTransitionUI and StateTransitionUITest to use updated log methods
