+++ src/main/kotlin/presentation/StateTransitionUI.kt
@@ -4,8 +4,8 @@ import org.baghdad.logic.model.entities.UserEntity
 import org.baghdad.logic.model.entities.UserType
 import org.baghdad.logic.model.exceptions.StateExceptions.NotFoundException
 import org.baghdad.logic.usecase.StateTransitionUseCase
-import org.baghdad.presentation.helper.Reader
-import org.example.presentation.Viewer
+import org.baghdad.presentation.input.Reader
+import org.baghdad.presentation.output.Viewer
 
 class StateTransitionUI(
     private val useCase: StateTransitionUseCase,
@@ -14,13 +14,13 @@ class StateTransitionUI(
 ) {
     fun execute() {
 
-        viewer.log("Enter Task ID:")
+        viewer.logMessage("Enter Task ID:")
         val taskId = reader.readInput()?.trim()
 
         if (taskId.isNullOrBlank()) throw Exception("Task ID cannot be null or blank.")
 
 
-        viewer.log("Enter New State ID:")
+        viewer.logMessage("Enter New State ID:")
         val newStateId = reader.readInput()?.trim()
 
         if (newStateId.isNullOrBlank()) throw Exception("New State ID cannot be null or blank.")
@@ -36,13 +36,13 @@ class StateTransitionUI(
         try {
             //todo
             useCase.changeTaskState(taskId, newStateId, user)
-            viewer.log("Task state changed successfully.")
+            viewer.logMessage("Task state changed successfully.")
         } catch (e: NotFoundException) {
-            viewer.log("State not found in this project: ${e.message}")
+            viewer.logError("State not found in this project: ${e.message}")
         } catch (e: IllegalStateException) {
-            viewer.log("Invalid operation: ${e.message}")
+            viewer.logError("Invalid operation: ${e.message}")
         } catch (e: Exception) {
-            viewer.log("Unexpected error: ${e.message}")
+            viewer.logError("Unexpected error: ${e.message}")
         }
     }
 }
